move method checkPageTitle to BaseMethods move 3 steps(with pageTitle) from ManageBrowser to WebPage add new comments for all steps classes
