+++ src/main/java/ru/alfabank/steps/BaseMethods.java
@@ -30,6 +30,7 @@ import io.restassured.specification.RequestSpecification;
 import lombok.SneakyThrows;
 import lombok.extern.slf4j.Slf4j;
 import org.apache.commons.io.FileUtils;
+import org.hamcrest.text.IsEqualIgnoringCase;
 import org.openqa.selenium.Keys;
 import ru.alfabank.alfatest.cucumber.api.AkitaScenario;
 import ru.alfabank.tests.core.rest.RequestParam;
@@ -46,6 +47,7 @@ import static com.codeborne.selenide.WebDriverRunner.getWebDriver;
 import static com.codeborne.selenide.WebDriverRunner.isIE;
 import static io.restassured.RestAssured.given;
 import static java.util.Objects.isNull;
+import static org.hamcrest.MatcherAssert.assertThat;
 import static org.junit.Assert.fail;
 import static ru.alfabank.alfatest.cucumber.ScopedVariables.resolveVars;
 import static ru.alfabank.tests.core.helpers.PropertyLoader.*;
@@ -324,7 +326,11 @@ public class BaseMethods {
         return Integer.parseInt(variableName.replaceAll("[^0-9]",""));
     }
 
-
-
+    public void checkPageTitle(String pageTitleName) {
+        pageTitleName = getPropertyOrStringVariableOrValue(pageTitleName);
+        String currentTitle = getWebDriver().getTitle().trim();
+        assertThat(String.format("Заголовок страницы не совпадает с ожидаемым значением. Ожидаемый результат: %s, текущий результат: %s", pageTitleName, currentTitle),
+                pageTitleName, IsEqualIgnoringCase.equalToIgnoringCase(currentTitle));
+    }
 
 }
