+++ src/main/java/ru/alfabank/steps/RoundUpSteps.java
@@ -1,3 +1,18 @@
+/**
+ * Copyright 2017 Alfa Laboratory
+ *
+ * Licensed under the Apache License, Version 2.0 (the "License");
+ * you may not use this file except in compliance with the License.
+ * You may obtain a copy of the License at
+ *
+ * http://www.apache.org/licenses/LICENSE-2.0
+ *
+ * Unless required by applicable law or agreed to in writing, software
+ * distributed under the License is distributed on an "AS IS" BASIS,
+ * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ * See the License for the specific language governing permissions and
+ * limitations under the License.
+ */
 package ru.alfabank.steps;
 
 import com.codeborne.selenide.Selenide;
@@ -30,7 +45,7 @@ import static ru.alfabank.tests.core.helpers.PropertyLoader.loadProperty;
 import static ru.alfabank.tests.core.helpers.PropertyLoader.loadValueFromFileOrPropertyOrVariableOrDefault;
 
 /**
- * Шаги для тестирования взаимодействия с внешним окружением (устройствами, файлами, property и т.д.)
+ * Шаги для тестирования взаимодействия с внешним окружением (устройствами, файлами, переменными окружения, property и т.д.)
  */
 
 @Slf4j
